Move footnotes to new file
